+++ packages/wordpress-kit/__mocks__/server/basicPostsQuery.ts
@@ -5,10 +5,13 @@ import basicPostsQueryData from '../../__tests__/data/basicPostsQuery.json';
 export const basicPostsQuery = graphql.query<typeof basicPostsQueryData>(
 	'BasicPostsQuery',
 	(req, res, ctx) => {
-		if (req.headers.has('Pantheon-SKey')) {
+		if (req.headers.has('Fastly-Debug')) {
 			return res(
 				ctx.data(basicPostsQueryData),
-				ctx.set('Surrogate-Key-Raw', 'post-7 post-1 user-1 graphql-collection'),
+				ctx.set(
+					'Surrogate-Key',
+					'oY1rHsX/3p6DpQK1dAe+ kGUbbVGynpd8ltN0W0Ce CShM/FbGRgcTR9hRB+ak Be0KFtdPtEV9mVlGh8Dq glzcfBFhDdGngbGRhqnD 2dajCGdEluE9JO0ZYeBT',
+				),
 			);
 		}
 		return res(ctx.data(basicPostsQueryData));