+++ packages/wordpress-kit/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @pantheon-systems/wordpress-kit
 
+## 2.15.0-canary.1
+
+### Minor Changes
+
+- ea3c25b2: The `GraphqlCLientFactory` now sets the `Fastly-Debug` header to
+  obtain hashed surrogate keys instead of raw surrogate keys
+
+### Patch Changes
+
+- Updated dependencies [b9a05706]
+  - @pantheon-systems/cms-kit@0.3.0-canary.0
+
+## 2.14.4-canary.0
+
+### Patch Changes
+
+- b9491fd0: Add tailwindcss as a project dependency.
+
 ## 2.14.3
 
 ### Patch Changes